+++ src/routes/test/edit/FileBar.svelte
