+++ src/routes/test/edit/ScriptPreview.svelte
