--- src/routes/test/playground/+page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import FileBar from './FileBar.svelte';
-	import Editor from './Editor.svelte';
-</script>
-
-<FileBar />
-<Editor />
